refactor(frontend): slice pattern in the store
