+++ packages/frontend/components/Borrow/CollateralDropdown.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { ListItemIcon, ListItemText, MenuItem, Typography } from "@mui/material"
 
-import { chains } from "../../store"
+import { chains } from "../../store/auth.slice"
 import Image from "next/image"
 
 type Chain = typeof chains[0]
